requirements.txt
python-telegram-bot==13.15
standard-imghdr==3.13.0
urllib3==1.26.15
environs~=14.5.0
redis~=7.1.0
vk_api~=11.10.0